Implements the Never `!` types as a TypeInfo bottom type.
We now parse the `!` as a TypeInfo::Never, and remove the usage of empty enums as Never type in our code.

This commit removes completely the DeterministicallyAborts and TypeCheckUnificationContext.

The DeterministicallyAborts can be removed because the Never TypeInfo is now propagated using the type checker.
Code blocks that return, break, continue, or call an expression that returns Never, are marked as Never.

Partially fixes #5562.