Fixes method application argument type checking.
While doing the argument type checking we set the type annotation to Unknown
 because the method has not been resolved yet.

With this change when we fail to type check an argument expression
with the annotation Unknown we try again after resolving the method name.

The reason we do the double pass on argument type checking is because we
need the types from the arguments to resolve the method but sometimes we
need the method resolved for resolving the argument expressions with the type parameters.