Adds reuse of enum variant TypeId.
We already do unification of expressions with type annotation when necessary,
but these were not affecting enum variants because we used separate TypeIds for
the expression and for the enum variant.

With this change the enum variant and the expressions will share the same TypeId,
making unification to the expression to also affect the enum variant.

Fixes #5492
Fixes #5581

Probably helpful for #5559