Implement Pid file locking
Fixes #5633

Added functionality to create and manage a lock file containing the process ID
(pid) of the running instance of the software. This mechanism prevents multiple
instances of the software from running simultaneously by checking the existence
and content of the lock file. If the lock file exists and contains a valid pid,
the struct will error gracefully to avoid conflicts. If the lock file is
missing or contains an invalid pid, the struct will proceed by removing the
file.  This ensures that only one instance of the software can run at a time
and it avoids stale locking to prevent future instances