Modify `LocationRange` to hold `LineCol` instead of absolute position.
This simplifies generating DWARF (as it uses line-col instead of pos),
and also simplifies some code in forc-debug launcher, where the
line-col information was being re-built from pos (and now doesn't need to).

On the other hand, some code in `addr2line` uses absolute position,
and I'm now rebuilding position from line-col information, rather
inefficiently in fact. This is okay as addr2line seems like a command
line utility that just does a reverse map from a code address to its
source line/col, so it doesn't need to be super-fast. In the future,
we could rewrite the code to directly use line-col information,
if required. It just seemed a bit complicated to do right now.

The DWARF information is written to `test.debug` at the moment and
can be verified using `llvm-dwarfdump --debug-line ./test.debug`.
We'll have to introduce a new flag or use the existing `-g` flag
(to write DWARF objects instead of JSONs).