Replace `__jmpb_ssp` with `__jmp_mem` (#5674)
The new intrinsic jumps to an absolute address that is fetched from
`MEM[$hp]`. This is more versatile than the `__jmpb_ssp` that was
previously provided. An immediate advantage being that we can have a
working version of `ldc` even in debug
builds.

The test has been modified too, to demonstrate an LDC use that works. A
working version is at
https://github.com/FuelLabs/ldc-testing/blob/309c79a8f73e26992123fb3ee9777b89dfda9ab1/test-contract/src/main.sw

Co-authored-by: Sophie Dankel <47993817+sdankel@users.noreply.github.com>