Fixes dummy method while using where clause in trait impl.
When using nested generic methods, as in the use case added, the lower method
would not be properly replaced and would remain as a trait dummy method.

The fix was to do find_method_for_type in the ReplaceDecls of FunctionAplication,
this allows the replacement of the function aplication dummy method with
a proper method implementation.

Fixes #5673