Fixes compiler asking for unnecessary disambiguation for trait methods.
`find_method_for_type` was throwing an error when multiple implementation
of a method exist receiving different kinds of unsigned int.
This was caused by the coercion comparison which returned true while
comparing any uint size to any other uint size.

The solution was to change the UnifyCheck on Coercion to only return true to same size uints.

Fixes #5798