ci: add a job for checking fuel-core versions between sdk-harness and test suite (#5930)
## Description
We are currently out of sync in between sdk-harness and test suite, this
is not ideal due to couple of reasons:

1. It is not nice to have our two test suites, testing with different
versions of fuel-core, which means we do not test one version completely
in the sense of our e2e and sdk-harness tests.
2. It can create problems in CI if a contract id needs to be updated,
when the out-of sync versions are incompatible. (for example: one is
using version 0.1.0, one is using 0.2.0)

The added job will fail and inform the dev bumping fuel-core version, to
also bump it in sdk-harness or in general make them in sync.

The reason we need this as a CI job is we cannot enforce it via cargo.
Since adding sdk-harness to the sway repo workspace, it creates our well
known cyclic dependency between sdk and sway. So this should be checked
in CI.