Return type of output_asset_to changed to Address (#5424)
## Description
When `output_asset_to` was added in #4694, it used `b256` as the return
type, instead of an `Address`. It referenced #4569, which describes the
confusion developers face between `Address` and `Identity`.

While that confusion _is_ an issue, I don't believe that's relevant to
this function type. An output coin can _only_ be an address, not a
contract, since contract balances are stored in the contract UTXO
itself. Furthermore, the standard lib API will need to be frozen soon,
and there's no plans in place to change anything in regards to #4569.

## Checklist

- [ ] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [ ] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.

Co-authored-by: SwayStar123 <46050679+SwayStar123@users.noreply.github.com>