Set code owners
We recently identified that our review process is a potential security
vulnerability, so we want to enforce that code changes are reviewed by
people who understand the code that has been changed, as well as enforce
that version bumps and releases are only done by a restricted group of
people.

This change is a first step towards that by setting the code owners to
reasonable values on the this repository.