Handle traits when doing namespace item imports.
This reverts PR #6080 due to a pretty severe performance regression.

The reason this was split to a separate step was because it needs a
typed declaration due to the depedency on the type system and thus
cannot be used during collection.

However, the type system will be changed to work with parsed
declarations pretty soon, so revert this back to the old behavior and
temporarily check for a typed declaration instead.

Fixes https://github.com/FuelLabs/sway/issues/6182.